chore(dev-next): restructure watchers
